changes after review 30.11.2020

+++ src/main/java/com/kodilla/ecommercee/repository/GroupRepository.java
@@ -8,6 +8,6 @@ import javax.transaction.Transactional;
 
 @Transactional
 @Repository
-public interface GroupRepository extends CrudRepository<GroupEntity, Integer> {
+public interface GroupRepository extends CrudRepository<GroupEntity, Long> {
 
 }
